Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-stereopanner-interface/stereopannernode-basic-expected.txt

CONSOLE MESSAGE: line 25: TypeError: context.createStereoPanner is not a function. (In 'context.createStereoPanner()', 'context.createStereoPanner' is undefined)

Harness Error (FAIL), message = TypeError: context.createStereoPanner is not a function. (In 'context.createStereoPanner()', 'context.createStereoPanner' is undefined)

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [test] Attributes and basic functionality of StereoPannerNode 
